requirements.txt
Django==3.2.5
face_recognition==1.3.0
keras==2.9.0
mediapipe==0.8.10
numpy==1.22.1
opencv_python==4.5.5.62
Pillow==9.1.1
tensorflow==2.9.1
XlsxWriter==3.0.3
